LOG-4743: Add constructor functions for source generator elements
