+++ internal/generator/vector/source/journal_log.go
@@ -12,3 +12,12 @@ journal_directory = "/var/log/journal"
 {{end}}`
 
 type JournalLog = framework.ConfLiteral
+
+func NewJournalLog(id string) JournalLog {
+	return JournalLog{
+		ComponentID:  id,
+		Desc:         "Logs from linux journal",
+		TemplateName: "inputSourceJournalTemplate",
+		TemplateStr:  JournalLogTemplate,
+	}
+}
