+++ internal/generator/vector/source/kubernetes_logs.go
@@ -1,7 +1,10 @@
 package source
 
 import (
+	"fmt"
+	"github.com/openshift/cluster-logging-operator/internal/constants"
 	"github.com/openshift/cluster-logging-operator/internal/generator/framework"
+	"strings"
 )
 
 type KubernetesLogs struct {
@@ -31,4 +34,44 @@ pod_annotation_fields.pod_node_name = "hostname"
 namespace_annotation_fields.namespace_uid = "kubernetes.namespace_id"
 rotate_wait_ms = 5000
 {{end}}`
+}
+
+func NewKubernetesLogs(id, namespace string) KubernetesLogs {
+	return KubernetesLogs{
+		ComponentID:  id,
+		Desc:         "Logs from containers (including openshift containers)",
+		ExcludePaths: ExcludeContainerPaths(namespace),
+	}
+}
+
+func LogFilesMetricExporterLogsPath(namespace string) string {
+	return fmt.Sprintf("/var/log/pods/%s_%%s-*/*/*.log", namespace)
+}
+
+func ElasticSearchLogStoreLogsPath(namespace string) string {
+	return fmt.Sprintf("/var/log/pods/%s_%%s-*/*/*.log", namespace)
+}
+
+func LokiLogStoreLogsPath(namespace string) string {
+	return fmt.Sprintf("/var/log/pods/%s_*/%%s*/*.log", namespace)
+}
+
+func VisualizationLogsPath(namespace string) string {
+	return fmt.Sprintf("/var/log/pods/%s_%%s-*/*/*.log", namespace)
+}
+
+func ExcludeContainerPaths(namespace string) string {
+	return fmt.Sprintf("[%s]", strings.Join(
+		[]string{
+			fmt.Sprintf("%q", fmt.Sprintf(LogFilesMetricExporterLogsPath(namespace), constants.LogfilesmetricexporterName)),
+			fmt.Sprintf("%q", fmt.Sprintf(ElasticSearchLogStoreLogsPath(namespace), constants.ElasticsearchName)),
+			fmt.Sprintf("%q", fmt.Sprintf(LokiLogStoreLogsPath(namespace), constants.LokiName)),
+			fmt.Sprintf("%q", fmt.Sprintf(VisualizationLogsPath(namespace), constants.KibanaName)),
+			fmt.Sprintf("%q", fmt.Sprintf("/var/log/pods/%s_*/%s/*.log", namespace, "gateway")),
+			fmt.Sprintf("%q", fmt.Sprintf("/var/log/pods/%s_*/%s/*.log", namespace, "opa")),
+			fmt.Sprintf("%q", "/var/log/pods/*/*/*.gz"),
+			fmt.Sprintf("%q", "/var/log/pods/*/*/*.tmp"),
+		},
+		", ",
+	))
 }
